+++ tests/unit/tests_mvd3.cpp
@@ -59,9 +59,9 @@ BOOST_AUTO_TEST_CASE( basicTestPosition )
     Positions neurons_pos = file.getPositions();
 
     BOOST_CHECK_EQUAL(neurons_pos.shape()[0], 1000);
-    BOOST_CHECK_EQUAL(neurons_pos[0][0], static_cast<double>(40.821401));
-    BOOST_CHECK_EQUAL(neurons_pos[0][1], static_cast<double>(1986.506637));
-    BOOST_CHECK_EQUAL(neurons_pos[0][2], static_cast<double>(10.788424));
+    BOOST_CHECK_EQUAL(neurons_pos[0][0], 40.821401);
+    BOOST_CHECK_EQUAL(neurons_pos[0][1], 1986.506637);
+    BOOST_CHECK_EQUAL(neurons_pos[0][2], 10.788424);
 
 }
 
@@ -330,9 +330,9 @@ BOOST_AUTO_TEST_CASE( basicExcFreq )
 
     std::vector<double> exc_frequencies = file.getExcMiniFrequencies();
 
-    BOOST_CHECK_EQUAL(exc_frequencies[0], static_cast<double>(0.63));
-    BOOST_CHECK_EQUAL(exc_frequencies[100], static_cast<double>(0.122));
-    BOOST_CHECK_EQUAL(exc_frequencies[200], static_cast<double>(0.04));
+    BOOST_CHECK_EQUAL(exc_frequencies[0], 0.63);
+    BOOST_CHECK_EQUAL(exc_frequencies[100], 0.122);
+    BOOST_CHECK_EQUAL(exc_frequencies[200], 0.04);
 
 }
 
@@ -345,9 +345,9 @@ BOOST_AUTO_TEST_CASE( basicInhFreq )
 
     std::vector<double> inh_frequencies = file.getInhMiniFrequencies();
 
-    BOOST_CHECK_EQUAL(inh_frequencies[0], static_cast<double>(0.012));
-    BOOST_CHECK_EQUAL(inh_frequencies[100], static_cast<double>(0.013));
-    BOOST_CHECK_EQUAL(inh_frequencies[200], static_cast<double>(0.014));
+    BOOST_CHECK_EQUAL(inh_frequencies[0], 0.012);
+    BOOST_CHECK_EQUAL(inh_frequencies[100], 0.013);
+    BOOST_CHECK_EQUAL(inh_frequencies[200], 0.014);
 
 }
 
@@ -398,9 +398,9 @@ BOOST_AUTO_TEST_CASE( genericMVDFile )
     Positions neurons_pos = file->getPositions();
 
     BOOST_CHECK_EQUAL(neurons_pos.shape()[0], 1000);
-    BOOST_CHECK_EQUAL(neurons_pos[0][0], static_cast<double>(40.821401));
-    BOOST_CHECK_EQUAL(neurons_pos[0][1], static_cast<double>(1986.506637));
-    BOOST_CHECK_EQUAL(neurons_pos[0][2], static_cast<double>(10.788424));
+    BOOST_CHECK_EQUAL(neurons_pos[0][0], 40.821401);
+    BOOST_CHECK_EQUAL(neurons_pos[0][1], 1986.506637);
+    BOOST_CHECK_EQUAL(neurons_pos[0][2], 10.788424);
 }
